src/general: consolidate write_raw_ascii + print_matrix into eigen_io.{h,cpp}
The Eigen-typed raw-ASCII writer and Matrix-print helpers were
duplicated across src/harmonic/main.cpp (from PR #184) and
src/atomic/inttest.cpp (from PR #185, this branch). Extract them
into helfem::io::{write_raw_ascii, print_matrix} in
src/general/eigen_io.{h,cpp} and use them from both call sites.

Adds `general/eigen_io.cpp` to the helfem-common source list so the
symbols are available to any FE test driver that wants the same
arma::mat::save / arma::mat::print output shape.
